Public API and better webpack compatibility   (#239)
* Generated declarations and exposed API

* Refactoring into more files
