+++ package.json
@@ -9,6 +9,8 @@
         "tstl",
         "transpiler"
     ],
+    "main": "dist/tstl.js",
+    "types": "dist/tstl.d.ts",
     "scripts": {
         "build": "tsc -p tsconfig.json && npm run build-lualib",
         "build-lualib": "ts-node ./build_lualib.ts",
@@ -41,7 +43,7 @@
         "node": ">=8.5.0"
     },
     "dependencies": {
-        "typescript": "^2.9.2",
+        "typescript": "2.9.2",
         "yargs": "^12.0.1"
     },
     "devDependencies": {
