+++ src/CompilerOptions.ts
@@ -1,0 +1,7 @@
+import * as ts from "typescript";
+
+export interface CompilerOptions extends ts.CompilerOptions {
+    addHeader?: boolean;
+    luaTarget?: string;
+    luaLibImport?: string;
+}
