+++ test/unit/cli.spec.ts
@@ -1,6 +1,6 @@
-import { Expect, Test, TestCase, Teardown } from "alsatian";
+import { Expect, Test, TestCase } from "alsatian";
 
-import { CompilerOptions, findConfigFile, parseCommandLine, ParsedCommandLine } from "../../src/CommandLineParser";
+import { findConfigFile, parseCommandLine } from "../../src/CommandLineParser";
 
 export class CLITests {
