+++ tsconfig.json
@@ -4,10 +4,11 @@
         "alwaysStrict" : true,
         "outDir": "./dist/",
         "rootDir": "./src/",
+        "declaration": true,
         "sourceMap": true,
         "experimentalDecorators": true,
         "target": "es6",
         "module": "commonjs"
     },
-    "exclude": ["test/**/*", "build_lualib.ts", "src/lualib"]
+    "exclude": ["test/**/*", "build_lualib.ts", "src/lualib", "dist/**/*"]
 }
